Implemented Facade Design Pattern for TileManager class
